Move `BufferBuilder` to `arrow-buffer`
